Add files via upload
As part of the Cisco Work experience we were given the task - in groups - to create a app which ensures an inclusive future for all individuals. Our group thought of the idea for an app called 'Balance' which helped individuals/adults who either needed some time to relax or had a mental health issue and wanted to find a safe environment to be able to discuss how they felt.